Beginning of time rewind mechanic
I have not finished this mechanic yet, but I have made some good headway. Sorry Adam, I've commandeered your test scene for the time being.
This feature requires timetravelenabler scripts to be assigned to enemies, and the timetravelSequence to be assigned to the player.
I've added some descriptions of the code changes I've made in those scripts